--- src/AddCardType.tsx
@@ -1,5 +1,0 @@
-const AddCardType = () => {
-  return <div>카드 선택 모달</div>
-}
-
-export default AddCardType
