--- src/components/CardTypeSelect.tsx
@@ -1,3 +1,0 @@
-const CardTypeSelect = () => {}
-
-export default CardTypeSelect
