+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './CardNumber'
+export * from './CardName'
+export * from './CardPassword'
+export * from './CardCvc'
+export * from './CardType'
+export * from './CardBox'
